backlight: Separate backlight level from exposed brightness
When we want the brightness curve to be non-linear be need two
values: those exposed to the user facing side and those values
actually set in the hardware (called `level`s).

We do the split here but use a linear mapping only. This will
change in the next commits.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1810>

+++ src/backlight.h
@@ -21,7 +21,7 @@ typedef enum {
 G_DECLARE_DERIVABLE_TYPE (PhoshBacklight, phosh_backlight, PHOSH, BACKLIGHT, GObject)
 
 int                 phosh_backlight_get_brightness (PhoshBacklight *self);
-void                phosh_backlight_set_brightness (PhoshBacklight *self, int brightness);
+void                phosh_backlight_set_brightness (PhoshBacklight *self, double brightness);
 double              phosh_backlight_get_relative (PhoshBacklight *self);
 void                phosh_backlight_set_relative (PhoshBacklight *self, double val);
 void                phosh_backlight_get_range (PhoshBacklight *self,
